requirements.txt
aniso8601==3.0.2
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.4
bcrypt==3.1.4
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
codeclimate-test-reporter==0.2.3
codecov==2.0.15
colorama==0.4.0
coverage==4.5.2
coveralls==1.5.1
docopt==0.6.2
Flask==1.0.2
Flask-JWT-Extended==3.13.1
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
idna==2.7
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
more-itertools==4.3.0
pbr==4.2.0
pipenv==2018.7.1
pluggy==0.8.0
psycopg2-binary==2.7.5
py==1.7.0
pycodestyle==2.4.0
pycparser==2.19
PyJWT==1.6.4
pylint==2.1.1
PyQt5==5.11.3
PyQt5-sip==4.19.13
pytest==4.0.1
pytest-cov==2.6.0
pytest-flask==0.13.0
python-coveralls==2.9.1
pytz==2018.5
PyYAML==3.13
requests==2.20.1
six==1.11.0
SQLAlchemy==1.2.12
stevedore==1.29.0
urllib3==1.24.1
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
Werkzeug==0.14.1
wrapt==1.10.11
